WIP: template rendering
- populate timestamp for testcase
